refactor(table): extract pure parse/serialize model into table-model.ts
extensions/table-widget.ts had grown to ~1,600 lines mixing four concerns
(pure model, widget/grid rendering, cell subview sync, keymap handlers). This
is the first, lowest-risk decomposition step from issue #570: move the pure
model layer (parseAligns, parseRowCells, parseTable, serializeRow,
serializeDelimiter, and the TableAlign/CellRange/TableData types) to a new
state-only extensions/table-model.ts. table-widget.ts imports from it;
table-model.test.ts now targets the new module. Move only — behavior
unchanged, full suite green.

Also documents the table keymap invariants (in-cell Tab/Enter/Shift-Enter/
arrow-escape, two-step Backspace select-then-delete, boundary-newline guard)
in packages/paper/docs/DECISIONS.md.
