+++ packages/paper/src/editor/extensions/table-widget.ts
@@ -15,6 +15,8 @@ import { Decoration, EditorView, ViewPlugin, WidgetType, keymap } from "@codemir
 import { renderCellInline } from "./cell-inline-renderer";
 import { guardedDecorations } from "./composing-guard";
 import { markdownExtension } from "./markdown";
+import type { TableAlign, TableData } from "./table-model";
+import { parseTable, serializeDelimiter, serializeRow } from "./table-model";
 
 // ADR-0002: A table is always a rendered grid widget; structure syntax (|, delimiter rows) stays permanently hidden.
 // ADR-0003: Editing happens in a CM subview shown only on the one focused cell — it shares the document with the
@@ -24,25 +26,7 @@ import { markdownExtension } from "./markdown";
 // Preventing table churn: even when cell editing changes the table source, the widget DOM is not recreated;
 // it's updated in place via updateDOM() — keeping the subview from being destroyed during editing is the key.
 
-export type TableAlign = "left" | "center" | "right" | null;
-
 export const cellSync = Annotation.define<"from-cell" | "from-main">();
-
-interface CellRange {
-  /** Cell content range in body document coordinates (trimmed). If editable=false, from/to are invalid */
-  from: number;
-  to: number;
-  text: string;
-  editable: boolean;
-}
-
-interface TableData {
-  source: string;
-  tableFrom: number;
-  aligns: TableAlign[];
-  /** rows[0] = header, rows[1..] = body rows (delimiter row excluded) */
-  rows: CellRange[][];
-}
 
 interface ActiveCell {
   tableFrom: number;
@@ -83,86 +67,6 @@ export function clearActiveCell(view: EditorView): void {
 /** Start cell editing on a specific table from the outside (e.g. a slash command) */
 export function activateCell(view: EditorView, tableFrom: number, row = 0, col = 0): void {
   view.dispatch({ effects: setActiveCellEffect.of({ tableFrom, row, col }) });
-}
-
-// ---------------------------------------------------------------------------
-// Parsing — use lezer only to locate the Table; compute cell ranges directly by scanning lines.
-// (lezer's TableCell node isn't created for empty cells, so it can't give a click entry position)
-
-export function parseAligns(delimiter: string): TableAlign[] {
-  return delimiter
-    .replace(/^\s*\|/, "")
-    .replace(/\|\s*$/, "")
-    .split("|")
-    .map((seg) => {
-      const s = seg.trim();
-      const left = s.startsWith(":");
-      const right = s.endsWith(":");
-      if (left && right) return "center";
-      if (right) return "right";
-      if (left) return "left";
-      return null;
-    });
-}
-
-export function parseRowCells(lineText: string, lineFrom: number): CellRange[] {
-  // Collect positions of unescaped pipes
-  const pipes: number[] = [];
-  for (let i = 0; i < lineText.length; i++) {
-    if (lineText[i] === "|" && lineText[i - 1] !== "\\") pipes.push(i);
-  }
-  if (pipes.length === 0) return [];
-
-  // GFM allows omitting the pipes at the row edges — compensate with virtual boundaries
-  const bounds = [...pipes];
-  if (lineText.slice(0, pipes[0]).trim() !== "") bounds.unshift(-1);
-  if (lineText.slice(pipes[pipes.length - 1] + 1).trim() !== "") bounds.push(lineText.length);
-
-  const cells: CellRange[] = [];
-  for (let i = 0; i < bounds.length - 1; i++) {
-    const segStart = bounds[i] + 1;
-    const segEnd = bounds[i + 1];
-    const seg = lineText.slice(segStart, segEnd);
-    // Treat only one padding space at each edge as outside the cell — a full trim would erase
-    // a trailing space the user just typed during the body→subview reverse sync
-    const leading = seg.startsWith(" ") ? 1 : 0;
-    const trailing = seg.length > leading && seg.endsWith(" ") ? 1 : 0;
-    const text = seg.slice(leading, seg.length - trailing);
-    const from = lineFrom + segStart + leading;
-    cells.push({ from, to: from + text.length, text, editable: true });
-  }
-  return cells;
-}
-
-const MISSING_CELL: CellRange = { from: -1, to: -1, text: "", editable: false };
-
-function parseTable(state: EditorState, nodeFrom: number, nodeTo: number): TableData {
-  const firstLine = state.doc.lineAt(nodeFrom);
-  const lastLine = state.doc.lineAt(nodeTo);
-  const rows: CellRange[][] = [];
-  let aligns: TableAlign[] = [];
-
-  for (let n = firstLine.number; n <= lastLine.number; n++) {
-    const line = state.doc.line(n);
-    if (n === firstLine.number + 1) {
-      aligns = parseAligns(line.text);
-      continue;
-    }
-    rows.push(parseRowCells(line.text, line.from));
-  }
-
-  const cols = Math.max(aligns.length, rows[0]?.length ?? 0);
-  for (const row of rows) {
-    while (row.length < cols) row.push(MISSING_CELL);
-    row.length = cols;
-  }
-
-  return {
-    source: state.sliceDoc(firstLine.from, lastLine.to),
-    tableFrom: firstLine.from,
-    aligns,
-    rows,
-  };
 }
 
 // ---------------------------------------------------------------------------
@@ -1020,25 +924,6 @@ interface ControllerHost extends HTMLElement {
 
 interface GripHost extends HTMLElement {
   __grip?: HTMLElement;
-}
-
-export function serializeRow(cells: string[]): string {
-  return "|" + cells.map((c) => (c ? ` ${c} ` : "  ")).join("|") + "|";
-}
-
-export function serializeDelimiter(aligns: TableAlign[]): string {
-  return (
-    "|" +
-    aligns
-      .map((a) => {
-        if (a === "center") return " :---: ";
-        if (a === "right") return " ---: ";
-        if (a === "left") return " :--- ";
-        return " --- ";
-      })
-      .join("|") +
-    "|"
-  );
 }
 
 // ---------------------------------------------------------------------------
